// src/api/generated/api.ts
export const ServerControllersModelsEnumsEnumCondition = {
    NUMBER_0: 0,
    NUMBER_1: 1,
    NUMBER_2: 2,
    NUMBER_3: 3,
    NUMBER_4: 4,
    NUMBER_5: 5,
    NUMBER_6: 6,
    NUMBER_7: 7,
    NUMBER_8: 8,
    NUMBER_9: 9,
    NUMBER_10: 10
} as const;

export type ServerControllersModelsEnumsEnumCondition = typeof ServerControllersModelsEnumsEnumCondition[keyof typeof ServerControllersModelsEnumsEnumCondition];



export const ServerControllersModelsEnumsEnumFlowers = {
    NUMBER_0: 0,
    NUMBER_1: 1,
    NUMBER_2: 2,
    NUMBER_3: 3,
    NUMBER_4: 4,
    NUMBER_5: 5,
    NUMBER_6: 6
} as const;

export type ServerControllersModelsEnumsEnumFlowers = typeof ServerControllersModelsEnumsEnumFlowers[keyof typeof ServerControllersModelsEnumsEnumFlowers];



export const ServerControllersModelsEnumsEnumFruit = {
    NUMBER_0: 0,
    NUMBER_1: 1,
    NUMBER_2: 2,
    NUMBER_3: 3,
    NUMBER_4: 4,
    NUMBER_5: 5,
    NUMBER_6: 6,
    NUMBER_7: 7,
    NUMBER_8: 8,
    NUMBER_9: 9,
    NUMBER_10: 10,
    NUMBER_11: 11,
    NUMBER_12: 12,
    NUMBER_13: 13,
    NUMBER_14: 14,
    NUMBER_15: 15,
    NUMBER_16: 16,
    NUMBER_17: 17,
    NUMBER_18: 18,
    NUMBER_19: 19
} as const;

export type ServerControllersModelsEnumsEnumFruit = typeof ServerControllersModelsEnumsEnumFruit[keyof typeof ServerControllersModelsEnumsEnumFruit];


/**
 * Требования к освещению для проращивания семян
 */

export const ServerControllersModelsEnumsEnumLight = {
    NUMBER_0: 0,
    NUMBER_1: 1,
    NUMBER_2: 2,
    NUMBER_3: 3,
    NUMBER_4: 4,
    NUMBER_5: 5,
    NUMBER_6: 6,
    NUMBER_7: 7,
    NUMBER_8: 8
} as const;

export type ServerControllersModelsEnumsEnumLight = typeof ServerControllersModelsEnumsEnumLight[keyof typeof ServerControllersModelsEnumsEnumLight];



export const ServerControllersModelsEnumsEnumReproduction = {
    NUMBER_0: 0,
    NUMBER_1: 1,
    NUMBER_2: 2,
    NUMBER_3: 3,
    NUMBER_4: 4,
    NUMBER_5: 5,
    NUMBER_6: 6,
    NUMBER_7: 7,
    NUMBER_8: 8,
    NUMBER_9: 9,
    NUMBER_10: 10,
    NUMBER_11: 11
} as const;

export type ServerControllersModelsEnumsEnumReproduction = typeof ServerControllersModelsEnumsEnumReproduction[keyof typeof ServerControllersModelsEnumsEnumReproduction];



export const ServerControllersModelsEnumsEnumViability = {
    NUMBER_0: 0,
    NUMBER_1: 1,
    NUMBER_2: 2,
    NUMBER_3: 3,
    NUMBER_4: 4,
    NUMBER_5: 5,
    NUMBER_6: 6
} as const;

export type ServerControllersModelsEnumsEnumViability = typeof ServerControllersModelsEnumsEnumViability[keyof typeof ServerControllersModelsEnumsEnumViability];


export interface ServerControllersModelsGrowthStageDTO {
    'id'?: string;
    'name'?: string | null;
    'description'?: string | null;
}
export interface ServerControllersModelsJournalRecordDTO {
    'id'?: string;
    'plantId'?: string;
    'growthStageId'?: string;
    'employeeId'?: string;
    'plantHeight'?: number;
    'fruitCount'?: number;
    'condition'?: ServerControllersModelsEnumsEnumCondition;
    'date'?: string;
}


export interface ServerControllersModelsLoginRequestDto {
    'username'?: string | null;
    'password'?: string | null;
}
export interface ServerControllersModelsLoginResponseDto {
    'username'?: string | null;
    'token'?: string | null;
}
export interface ServerControllersModelsPlantDTO {
    'id'?: string;
    'clientId'?: string;
    'specie'?: string | null;
    'family'?: string | null;
    'flower'?: ServerControllersModelsEnumsEnumFlowers;
    'fruit'?: ServerControllersModelsEnumsEnumFruit;
    'reproduction'?: ServerControllersModelsEnumsEnumReproduction;
}


export interface ServerControllersModelsRegisterRequestDto {
    'email'?: string | null;
    'password'?: string | null;
}
export interface ServerControllersModelsSeedDTO {
    'id'?: string;
    'plantId'?: string;
    'maturity'?: string | null;
    'viability'?: ServerControllersModelsEnumsEnumViability;
    'lightRequirements'?: ServerControllersModelsEnumsEnumLight;
    'waterRequirements'?: string | null;
    'temperatureRequirements'?: number;
}